refactor: funciona
